requirements.txt
requests==2.26.0
lxml==4.6.4
beautifulsoup4==4.10.0
google-cloud-storage==1.43.0
google-cloud-bigquery==2.31.0
